requirements.txt
requests>=2.25.0
aiohttp>=3.7.0
tqdm>=4.0.0
psutil>=5.8.0

colorama>=0.4.4
prompt-toolkit>=3.0.0

pyrogram>=2.0.0
tgcrypto>=1.2.0

pymongo>=4.0.0
motor>=2.5.0

python-dotenv>=0.19.0
loguru>=0.6.0

pydantic-settings>=2.7.1
pydantic>=2.9.2